% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AUC}
\alias{AUC}
\title{Negative AUC for matrix predictions}
\usage{
AUC(A, P)
}
\arguments{
\item{A}{Observed binary adjacency (matrix or vector).}

\item{P}{Predicted probability (matrix or vector).}
}
\value{
Negative AUC value (for minimisation-based model selection).
}
\description{
Negative AUC for matrix predictions
}
