% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auroc}
\alias{auroc}
\title{AUC via Wilcoxon statistic}
\usage{
auroc(score, bool)
}
\arguments{
\item{score}{Numeric prediction scores.}

\item{bool}{Logical/binary label vector.}
}
\value{
AUC value.
}
\description{
AUC via Wilcoxon statistic
}
\keyword{internal}
