% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin.dev}
\alias{bin.dev}
\title{Binomial deviance loss}
\usage{
bin.dev(x, y)
}
\arguments{
\item{x}{Observed binary matrix/vector.}

\item{y}{Predicted probability matrix/vector.}
}
\value{
Total binomial deviance (non-finite terms set to 0).
}
\description{
Binomial deviance loss
}
