% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{blockmodel.gen.fast}
\alias{blockmodel.gen.fast}
\title{Generate a fast SBM or DCBM network (sparse)}
\usage{
blockmodel.gen.fast(n, K, B, psi = rep(1, n), PI = rep(1/K, K), ncore = 1)
}
\arguments{
\item{n}{Number of nodes.}

\item{K}{Number of communities.}

\item{B}{K x K block probability matrix.}

\item{psi}{Degree heterogeneity vector (length \code{n}); set to all 1s for SBM.}

\item{PI}{Community prior probabilities (length \code{K}).}

\item{ncore}{Number of cores for parallel computation.}
}
\value{
A list with:
  \item{A}{Sparse symmetric adjacency matrix.}
  \item{member}{Community membership vector.}
  \item{psi}{Scaled degree heterogeneity vector.}
}
\description{
Uses parallel edge sampling for speed; returns a sparse adjacency matrix.
}
