% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissant.R
\name{croissant.latent}
\alias{croissant.latent}
\title{CROISSANT for latent space model dimension selection}
\usage{
croissant.latent(
  A,
  d.cand,
  s,
  o,
  R,
  loss = c("l2", "bin.dev", "AUC"),
  ncore = 1
)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{d.cand}{Candidate embedding ranks.}

\item{s}{Number of non-overlapping subsamples.}

\item{o}{Overlap size.}

\item{R}{Number of repetitions.}

\item{loss}{Loss functions to evaluate.}

\item{ncore}{Number of cores.}
}
\value{
A list with loss table and selected dimension per loss.
}
\description{
Selects the latent dimension for a latent space network model using
the CROISSANT framework with MLE fitting.
}
