% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{eigen.DCBM.est}
\alias{eigen.DCBM.est}
\title{Eigenvector-based DCBM estimation}
\usage{
eigen.DCBM.est(A, g, n = nrow(A), K = max(g), psi.omit = 0, p.sample = 1)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{g}{Community label vector.}

\item{n}{Number of nodes.}

\item{K}{Number of communities.}

\item{psi.omit}{Number of leading nodes to exclude from psi estimation
(used in CROISSANT overlapping designs).}

\item{p.sample}{Sampling proportion for correction (default 1).}
}
\value{
A list with \code{Bsum} and \code{psi}.
}
\description{
Estimates DCBM parameters using the row-norms of the leading eigenvectors
for degree heterogeneity.
}
