% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fast.DCBM.est}
\alias{fast.DCBM.est}
\title{Fast DCBM parameter estimation}
\usage{
fast.DCBM.est(A, g, n = nrow(A), K = max(g), psi.omit = 0, p.sample = 1)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{g}{Community label vector.}

\item{n}{Number of nodes.}

\item{K}{Number of communities.}

\item{psi.omit}{Number of leading nodes to exclude from psi estimation
(used in CROISSANT overlapping designs).}

\item{p.sample}{Sampling proportion for correction (default 1).}
}
\value{
A list with:
  \item{Bsum}{K x K block sum matrix (divided by \code{p.sample}).}
  \item{psi}{Degree heterogeneity vector.}
}
\description{
Estimates the block sum matrix and degree heterogeneity parameters
for a Degree-Corrected Block Model.
}
