% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fast.SBM.est}
\alias{fast.SBM.est}
\title{Fast SBM block probability estimation}
\usage{
fast.SBM.est(A, g, n = nrow(A), K = max(g))
}
\arguments{
\item{A}{Adjacency matrix (n x n).}

\item{g}{Integer community label vector (length n).}

\item{n}{Number of nodes.}

\item{K}{Number of communities.}
}
\value{
K x K estimated block probability matrix.
}
\description{
Estimates the K x K block probability matrix from an adjacency matrix
and community labels.
}
