% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l2}
\alias{l2}
\title{L2 loss between two matrices}
\usage{
l2(x, y)
}
\arguments{
\item{x, y}{Numeric matrices of the same dimension.}
}
\value{
The Frobenius norm \code{sqrt(sum((x - y)^2))}.
}
\description{
L2 loss between two matrices
}
