% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{latent.gen}
\alias{latent.gen}
\title{Generate a latent space network}
\usage{
latent.gen(n, d, alpha = 1, sparsity = 1, ncore = 1)
}
\arguments{
\item{n}{Number of nodes.}

\item{d}{Latent dimension.}

\item{alpha}{Intercept parameter controlling overall density.}

\item{sparsity}{Sparsity multiplier.}

\item{ncore}{Number of cores for parallel computation.}
}
\value{
A list with:
  \item{A}{Sparse symmetric adjacency matrix.}
  \item{Z}{n x d latent position matrix.}
}
\description{
Generate a latent space network
}
