% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{matched.lab}
\alias{matched.lab}
\title{Apply label permutation to match reference}
\usage{
matched.lab(lab, fixed, n = length(lab), K = max(lab, fixed))
}
\arguments{
\item{lab}{Integer vector of labels to relabel.}

\item{fixed}{Integer vector of reference labels.}

\item{n}{Number of nodes.}

\item{K}{Number of communities.}
}
\value{
Relabelled integer vector aligned to \code{fixed}.
}
\description{
Apply label permutation to match reference
}
