% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neglog}
\alias{neglog}
\title{Safe negative log-likelihood term}
\usage{
neglog(n = 1, p = 0.5)
}
\arguments{
\item{n}{Numeric count.}

\item{p}{Probability (between 0 and 1).}
}
\value{
Numeric value of \code{-n * log(p)} or 0 if \code{p <= 0}.
}
\description{
Computes \code{-n * log(p)}, returning 0 when \code{p <= 0}.
}
