% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sparse.RDPG.gen}
\alias{sparse.RDPG.gen}
\title{Generate a sparse RDPG network}
\usage{
sparse.RDPG.gen(n, d, sparsity.multiplier = 1, ncore = 1)
}
\arguments{
\item{n}{Number of nodes.}

\item{d}{Latent dimension.}

\item{sparsity.multiplier}{Multiplier to control network density.}

\item{ncore}{Number of cores for parallel computation.}
}
\value{
A list with:
  \item{A}{Sparse symmetric adjacency matrix.}
  \item{P}{n x n probability matrix.}
}
\description{
Generate a sparse RDPG network
}
