% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumFast}
\alias{sumFast}
\title{Fast sum of matrix or vector elements}
\usage{
sumFast(X)
}
\arguments{
\item{X}{A matrix or vector.}
}
\value{
The total sum of all elements.
}
\description{
Fast sum of matrix or vector elements
}
\keyword{internal}
